REM-40 feat: обновление компонента FileNode для поддержки промежуточного состояния чекбокса и улучшение тестов
